vectorizers use ints, tf remote model creates int32 protos

+++ python/baseline/vectorizers.py
@@ -84,7 +84,7 @@ class Token1DVectorizer(AbstractVectorizer):
         if self.mxlen < 0:
             self.mxlen = self.max_seen
 
-        vec1d = np.zeros(self.mxlen, dtype=np.int32)
+        vec1d = np.zeros(self.mxlen, dtype=int)
         for i, atom in enumerate(self._next_element(tokens, vocab)):
             if i == self.mxlen:
                 i -= 1
@@ -195,7 +195,7 @@ class Char2DVectorizer(AbstractCharVectorizer):
 
         EOW = vocab.get('<EOW>', vocab.get(' '))
 
-        vec2d = np.zeros((self.mxlen, self.mxwlen), dtype=np.int32)
+        vec2d = np.zeros((self.mxlen, self.mxwlen), dtype=int)
         i = 0
         j = 0
         for atom in self._next_element(tokens, vocab):
@@ -257,7 +257,7 @@ class Char1DVectorizer(AbstractCharVectorizer):
         if self.mxlen < 0:
             self.mxlen = self.max_seen_tok
 
-        vec1d = np.zeros(self.mxlen, dtype=np.int32)
+        vec1d = np.zeros(self.mxlen, dtype=int)
         for i, atom in enumerate(self._next_element(tokens, vocab)):
             if i == self.mxlen:
                 i -= 1
